Fully adjusted move to shared
