+++ src/OpenTelemetry.ResourceDetectors.AWS/AWSResourcesEventSource.cs
@@ -26,10 +26,4 @@ internal sealed class AWSResourcesEventSource : EventSource
     {
         this.WriteEvent(3, format, exception);
     }
-
-    [Event(2, Message = "Failed to validate certificate in format: '{0}', error: '{1}'.", Level = EventLevel.Warning)]
-    public void FailedToValidateCertificate(string format, string error)
-    {
-        this.WriteEvent(4, format, error);
-    }
 }
